Minor fixes

+++ src/sdl_utils.cpp
@@ -73,7 +73,7 @@ void render_display(Chip8 &chip8, SDL_Renderer **renderer, const std::uint32_t &
     {
         for (std::uint32_t y{0}; y < WINDOW_HEIGHT; y++)
         {
-            std::uint8_t pixel_value = chip8.display.at(x + y * WINDOW_WIDTH);
+            std::uint32_t pixel_value = chip8.display.at(x + y * WINDOW_WIDTH);
 
             // Uses pixel_value for RGB as it should always be either 0x00 or 0xFF
             SDL_SetRenderDrawColor(*renderer, pixel_value, pixel_value, pixel_value, 0xFF);
